requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
playwright==1.40.0
PyYAML==6.0.1
python-dotenv==1.0.0
google-cloud-storage==2.13.0
google-cloud-secret-manager==2.16.4
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-django==0.42b0
gunicorn==21.2.0
pytest==7.4.3
pytest-django==4.7.0
